refactor: 페이지 별로 파일 정리하기
