+++ src/App.tsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import Footer from "./Footer";
 import Header from "./Header";
-import Main from "./Main";
-import ThemePage from "./ThemePage";
+import Main from "./MainPage/Main";
+import ThemePage from "./ThemePage/Theme";
 
 const App: React.FC = () => {
   //const name = 'Josh Perez';
