+++ src/MainPage/GoodsList.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { useState } from 'react';
 
-import { RankingGoodsItems } from './components/common/GoodsItem/Ranking';
+import { RankingGoodsItems } from '../components/common/GoodsItem/Ranking';
 
 const GoodsList = () => {
   const items = [
